pyproject.toml
[project]
    authors=[{name="Chimera"}]
    dependencies=[
        "httpx>=0.27",
        "inflection>=0.5",
        "jinja2>=3.1",
        "pydantic>=2.8",
        "pytest-asyncio>=0.23",
        "pytest>=8.3",
        "python-slugify>=8.0",
        "ruff>=0.5",
        "typer>=0.12",
    ]
    description="Async-only, type-safe Python client generator for InstantDB"
    name="instant-client"
    requires-python=">=3.11"
    version="0.2.0"

    [project.scripts]
        instant="instant_client.cli:app"

[build-system]
    build-backend="uv_build"
    requires=["uv_build>=0.8.5,<0.9.0"]


[tool.ruff]
    line-length=100
    select=["E", "F", "I", "UP"]
